Mostly transport operator tweaks
Converted the transport operator to an interface with one class each for
surface and delta tracking. Removed some of the cumbersome calls to
coords in transport operators by adding coord functions into particle.
In the process of writing a purely transformation-based boundary
condition procedure to delta tracking to resolve the long standing bug.