Modified geometry to identify and label unique cell instances
Still a big error present that needs to be found - but it compiles!
Added a countCell and locateCell routine to geometry: the former
identifies the number of instances, the latter obtains their coordList.
These are used to create cell instances. Also modified
initialiseGeometry to account for this, making the material fill the
final step of the initialisation. Also cleaned it up a bit - no need to
return cells, universes etc when they are contained in the geometry.
Still need to do testing.