Created new clases to store results. Will replace the tallyCounter
New module tallyEstimator was created. It contains clases: 
tallyScore    -> stores single score (only sum of scores) 
tallyCounter  -> stores sum of scores and scores^2. Calculates estimate
tallyEstimator-> composition of tallyScore and tallyCounter. Can
accumulate scores in a single history or cycle untill a batch is closed
and history or cycle-wise estimate is accumulated in tallyCounter. This
will be used for estimates that are composed from a single integration.