Updated tallyMapSlot and tallyMapFactory.
There is a new pattern for dealing with a polymorphic collection of
objects with a common interface, slot and factory, which avoids memory
leaks. Basically factory procedures are now subroutines. In tallyMaps
factory is also hidden behind slot init procedure, which is also part of
the abstract interface. The last think may change.