Created ceNeutronNuclide Interface
Design converged on a bit weird structure where the nuclide needs to
store a pointer to a database in which it is stored in order to ask for
an update of XSs on the cache. 

This structure allows to easily expose internal data of a nuclide if
this is required while also allowing interpolation of XSs to be
performed at a database level so it is possible to take advantage of
extra data interpolation if it is convenient.