Fix ignored fission for exotic nuclides
For a number of exotic nuclides (e.g. Am242 in ground state) prompt
fission spectra are given for Nth-chance fissions (MT=19,20,21,38). This
causes NJOY to not produce FIS block and MT=18 in ACE card. This commit
fixes a bug, which caused fission to be ignored in such a case.

Now, ACE nuclide is determined fissile if JXS(2)/=0 (has NU data). Then, if
FIS block is missing, MT=18 reaction is reconstructed as a sum of
Nth-chance fissions. Prompt spectrum is taken from MT=19 (n,f) reaction.