Allow offset to be disabled in nested universes
It is desirable for input to allow some universes not to be offset as
might be done in a lattice. ALternatively, it may be preferable to
evaluate their locations in the global reference frame. The former has
been done by allowing lattice universes to have an 'offset map', the
latter has been done by allowing a chosen universe to be evaluated in
global coordinates. This allowed a fix to the BEAVRS inputs to correctly
position the instrumentation tubes.

There seems to be a dictionary error with the offsetMap. Occasionally it
is not recognised as an integer list. This has prevented the creation of
a new test in latUniverse. This has been commented out until the
dictionary has been fixed.