Discontinuous field (#171)
* Fix to keff for MG scattering

* BEAVRS_HZP and geometry fixes

Added in the BEAVRS model with the D-bank partially inserted. Also fixed
a flaw in the geometry where the outermost cell was defined such that
there could be a rare particle lost between it and the geometry
boundary.

* Added distance calculation to some tallyMaps

* Added piecewise constant fields

Field which have values that are piecewise constant. Endowed with a
distance calculation. Added a simple version which allows a lattice to
be specified.

* PieceConstant fields take coordList

* Silly BEAVRS fix...

* Fixes to silly things

* Field interface accepts particles or coords

* Updating field interface

* Addressing PR comments